requirements.txt
pillow>=10.0
pytesseract>=0.3.10
numpy>=1.25
pytest>=7.4
